Working on the profileScreen layout.
