+++ BaseUI/src/components/BackgroundForm.tsx
@@ -16,20 +16,23 @@ class BackgroundForm extends React.Component<BackgroundFormProps, {}> {
         <ImageBackground
           source={backgroundImage}
           style={BackgroundFormStyles.backgroundImageStyle}>
-          <View style={BackgroundFormStyles.profileHeaderText}>
-            {this.props?.mainTitleElement}
-            {this.props?.editButtonElement &&
-              !this.props?.editMode &&
-              this.props.editButtonElement}
-          </View>
-          <View style={BackgroundFormStyles.profileImageContainer}>
-            {this.props?.profileImageElement}
-          </View>
+          {/* <View style={BackgroundFormStyles.profileHeaderText}> */}
+          {this.props?.mainTitleElement}
+          {this.props?.editButtonElement &&
+            !this.props?.editMode &&
+            this.props.editButtonElement}
+          {/* </View> */}
+          {/* <View style={BackgroundFormStyles.profileImageContainer}> */}
+          {this.props?.profileImageElement}
+          {/* </View> */}
           <KeyboardAvoidingView
-            behavior={Platform.OS === 'ios' ? 'padding' : undefined}>
-            <View style={BackgroundFormStyles.viewStyle}>
-              {this.props.children}
-            </View>
+            style={{height: '80%', borderColor: 'yellow', borderWidth: 2}}>
+            <KeyboardAvoidingView
+              behavior={Platform.OS === 'ios' ? 'padding' : undefined}>
+              <View style={BackgroundFormStyles.viewStyle}>
+                {this.props.children}
+              </View>
+            </KeyboardAvoidingView>
           </KeyboardAvoidingView>
         </ImageBackground>
       </>
