+++ BaseUI/src/components/SocialSection.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import {StyleSheet, View} from 'react-native';
+import {View} from 'react-native';
 import SocialSectionElement from './SocialSectionElement';
+import SocialSectionStyles from '../styles/SocialSectionStyles';
+import SocialSectionElementStyles from '../styles/SocialSectionElementStyles';
+import SocialSectionProps from '../interfaces/SocialSectionProps';
 
-class SocialSection extends React.Component<
-  {followers: number; following: number},
-  {}
-> {
+class SocialSection extends React.Component<SocialSectionProps, {}> {
   render(): React.ReactNode {
     return (
       <View style={SocialSectionStyles.containerStyle}>
@@ -16,7 +16,9 @@ class SocialSection extends React.Component<
           numberTextStyle={SocialSectionElementStyles.numberTextStyle}
           titleTextStyle={SocialSectionElementStyles.titleTextStyle}
         />
+
         <View style={SocialSectionStyles.separatorStyle} />
+
         <SocialSectionElement
           numberText={this.props.following}
           titleText={'Followings'}
@@ -28,37 +30,5 @@ class SocialSection extends React.Component<
     );
   }
 }
-
-const SocialSectionElementStyles = StyleSheet.create({
-  containerStyle: {
-    width: '50%',
-    flexDirection: 'column',
-    //paddingTop: 22,
-    paddingVertical: 22,
-    //borderWidth: 1,
-  },
-  numberTextStyle: {color: 'red', textAlign: 'center', fontSize: 32},
-  titleTextStyle: {color: 'gray', textAlign: 'center'},
-});
-
-const SocialSectionStyles = StyleSheet.create({
-  containerStyle: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    width: '80%',
-    borderWidth: 0.5,
-    borderColor: 'rgb(181, 182, 221)',
-    borderRadius: 20,
-    marginTop: 42,
-    //borderWidth: 1,
-  },
-  separatorStyle: {
-    height: '40%',
-    borderWidth: 0.5,
-    borderColor: 'rgb(181, 182, 221)',
-    alignSelf: 'center',
-    //marginTop: 22,
-  },
-});
 
 export default SocialSection;
