+++ BaseUI/src/interfaces/SocialSectionElementProps.ts
@@ -1,0 +1,14 @@
+import {
+  TextStyle,
+  ViewStyle,
+} from 'react-native/Libraries/StyleSheet/StyleSheetTypes';
+
+interface SocialSectionElementProps {
+  numberText: number;
+  titleText: string;
+  containerStyle: ViewStyle;
+  numberTextStyle: TextStyle;
+  titleTextStyle: TextStyle;
+}
+
+export default SocialSectionElementProps;
