+++ BaseUI/src/interfaces/SocialSectionProps.ts
@@ -1,0 +1,6 @@
+interface SocialSectionProps {
+  followers: number;
+  following: number;
+}
+
+export default SocialSectionProps;
