+++ BaseUI/src/styles/SocialSectionElementStyles.ts
@@ -1,0 +1,15 @@
+import {StyleSheet} from 'react-native';
+
+const SocialSectionElementStyles = StyleSheet.create({
+  containerStyle: {
+    width: '50%',
+    flexDirection: 'column',
+    //paddingTop: 22,
+    paddingVertical: 22,
+    //borderWidth: 1,
+  },
+  numberTextStyle: {color: 'red', textAlign: 'center', fontSize: 32},
+  titleTextStyle: {color: 'gray', textAlign: 'center'},
+});
+
+export default SocialSectionElementStyles;
