+++ BaseUI/src/styles/SocialSectionStyles.ts
@@ -1,0 +1,23 @@
+import {StyleSheet} from 'react-native';
+
+const SocialSectionStyles = StyleSheet.create({
+  containerStyle: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    width: '80%',
+    borderWidth: 0.5,
+    borderColor: 'rgb(181, 182, 221)',
+    borderRadius: 20,
+    marginTop: 42,
+    //borderWidth: 1,
+  },
+  separatorStyle: {
+    height: '40%',
+    borderWidth: 0.5,
+    borderColor: 'rgb(181, 182, 221)',
+    alignSelf: 'center',
+    //marginTop: 22,
+  },
+});
+
+export default SocialSectionStyles;
